fix(doc): remove line # from README link
The `README.md` link to the file that defines which version of Charist will be used was not pointing to the most relevant line number in that file. Since this documentation is not automatically generated it is likely that as lines are added or removed from this file this line number will change again. Therefore it has been removed. The file is sufficiently short that this shouldn't cause any trouble.

Closes #68